changes to make both auth work
