+++ uli_community/lib/uli_community_web/controllers/user_session_controller.ex
@@ -4,33 +4,20 @@ defmodule UliCommunityWeb.UserSessionController do
   alias UliCommunity.Accounts
   alias UliCommunityWeb.UserAuth
 
-  def create(conn, %{"_action" => "registered"} = params) do
-    create(conn, params, "Account created successfully!")
+  def new(conn, _params) do
+    render(conn, :new, error_message: nil)
   end
 
-  def create(conn, %{"_action" => "password_updated"} = params) do
-    conn
-    |> put_session(:user_return_to, ~p"/users/settings")
-    |> create(params, "Password updated successfully!")
-  end
-
-  def create(conn, params) do
-    create(conn, params, "Welcome back!")
-  end
-
-  defp create(conn, %{"user" => user_params}, info) do
+  def create(conn, %{"user" => user_params}) do
     %{"email" => email, "password" => password} = user_params
 
     if user = Accounts.get_user_by_email_and_password(email, password) do
       conn
-      |> put_flash(:info, info)
+      |> put_flash(:info, "Welcome back!")
       |> UserAuth.log_in_user(user, user_params)
     else
       # In order to prevent user enumeration attacks, don't disclose whether the email is registered.
-      conn
-      |> put_flash(:error, "Invalid email or password")
-      |> put_flash(:email, String.slice(email, 0, 160))
-      |> redirect(to: ~p"/users/log_in")
+      render(conn, :new, error_message: "Invalid email or password")
     end
   end
